Move autocanary.py logic into __init__.py
